pyproject.toml
[tool.poetry]
name = "mnmnwag"
version = "3.1"
description = ""
authors = ["Matthew Newton <matthewn@alum.berkeley.edu>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2"
wagtail = "*"
django-dbbackup = "^3.3.0"
django-debugtools = "*"
django-extensions = "*"
werkzeug = "*"
python-dateutil = "*"
django-compressor = "*"
django-libsass = "*"
django-capture-tag = "*"
django-comments-xtd = "*"
django-debug-toolbar = "*"
django-markdown2 = "*"
psycopg2 = "*"
wagtailmedia = "*"
tweepy = "*"
django-cachalot = "*"
django-browser-reload = "*"
pywatchman = "*"
django-honeypot = "*"
pyexiv2 = "^2.7.1"
gunicorn = "^20.1.0"
setproctitle = "^1.2.2"
seevooplay = "^1.0.5"
django-antispam = "^1.0.3"
django-crequest = "^2018.5.11"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
